For work

+++ Inventary.Services/Services/ItemService.cs
@@ -35,7 +35,7 @@ public class ItemService : IItemService
 
     public async Task<IList<ItemsList>> GetItemsListAsync()
     {
-        return await _repositoryManager.ItemRepository.GetListItemsAsync();
+        return await _repositoryManager.ItemRepository.GetListItemsWithoutSetupAsync();
     }
 
     public async Task<IList<ItemsList>> GetItemsListBySetupId(Guid id)
@@ -55,28 +55,31 @@ public class ItemService : IItemService
 
     public async Task<ItemDto> CreateAsync(CreateItemDto createItem)
     {
-        var item = _mapper.Map<Item>(createItem);
+
+        var newItem = _mapper.Map<Item>(createItem);
+
+        var item = await _repositoryManager.ItemRepository.Add(newItem);
 
         var result = _mapper.Map<ItemDto>(item);
-
-        await _repositoryManager.ItemRepository.Add(item);
 
         await _repositoryManager.UnitOfWork.SaveChangesAsync();
         return result;
     }
 
-    public async Task<ItemDto> UpdateAsync(Guid id, UpdateItemDto item)
+    public async Task<ItemDto> UpdateAsync(Guid itemId, UpdateItemDto item)
     {
-        var desiredItem = await _repositoryManager.ItemRepository.GetByIdAsync(id);
+        var notNullableUserDate = item.UserDate ?? DateTime.UtcNow;
+
+        var desiredItem = await _repositoryManager.ItemRepository.GetByIdAsync(itemId);
         if (desiredItem is null)
-            throw new ItemNotFoundException(id);
+            throw new ItemNotFoundException(itemId);
         var mappedAttachments = _mapper.Map<List<Attachment>>(item.Attachments);
         var mappedItemPhotos = _mapper.Map<List<ItemPhoto>>(item.ItemPhotos);
         var mappedDefects = _mapper.Map<List<Defect>>(item.Defects);
         var mappedComments = _mapper.Map<List<Comment>>(item.Comments);
         desiredItem.UpdateDate = DateTime.UtcNow;
         desiredItem.ItemName = item.ItemName;
-        desiredItem.UserDate = item.UserDate;
+        desiredItem.UserDate = notNullableUserDate;
         desiredItem.Status = item.Status;
         desiredItem.Price = item.Price;
         desiredItem.QRcode = item.QRcode;
@@ -88,7 +91,7 @@ public class ItemService : IItemService
         desiredItem.Attachments = mappedAttachments;
         desiredItem.Defects = mappedDefects;
         desiredItem.Comments = mappedComments;
-        
+
         // await _repositoryManager.ItemRepository.Update(deciredItem);
         await _repositoryManager.UnitOfWork.SaveChangesAsync();
 
@@ -96,30 +99,36 @@ public class ItemService : IItemService
         return result;
     }
 
-    public async Task MoveItemsToAnotherRoom(Guid id, IList<ListItemsForUpdate> items)
+    public async Task<bool> MoveItemsToAnotherRoom(Guid roomId, IList<ListItemsForUpdate> items)
     {
-        var desiredRoom = await _repositoryManager.RoomRepository.GetByIdAsync(id);
+        var desiredRoom = await _repositoryManager.RoomRepository.GetByIdAsync(roomId);
         if (desiredRoom is null)
-            throw new RoomNotFoundException(id);
+            throw new RoomNotFoundException(roomId);
 
         var itemsList = items.ToList();
-        // Можно ли это оптимизировать? Мб отправлять массив на апдейт? И как тогда менять значения...
-        if (itemsList is not null)
+        if (itemsList is null || itemsList.Count == 0)
+        {
+            throw new Exception("List items cannot be empty!");
+        }
+        else
         {
             foreach (var item in itemsList)
             {
                 var findItem = await _repositoryManager.ItemRepository.GetByIdAsync(item.Id);
                 if (findItem is null) continue;
                 findItem.UpdateDate = DateTime.UtcNow;
-                findItem.RoomId = id;
+                findItem.RoomId = roomId;
                 findItem.SetupId = null;
                 findItem.Setup = null;
                 findItem.Status = StatusEnum.StatusType.Active;
                 await _repositoryManager.UnitOfWork.SaveChangesAsync();
             }
+
+            return true;
         }
+        
     }
-    
+
     public async Task<ItemDto> DeleteAsync(Guid id)
     {
         var item = await _repositoryManager.ItemRepository.GetByIdAsync(id);
@@ -132,11 +141,17 @@ public class ItemService : IItemService
         var result = _mapper.Map<ItemDto>(item);
         return result;
     }
-    
-    public async Task DeleteRange(IList<ItemsForRoom> items)
+
+    public async Task<bool> DeleteRange(IList<ItemsForRoom> items)
     {
+        if (items is null || items.Count == 0)
+        {
+            throw new Exception("List items cannot be empty.");
+        }
         var mappedListItems = _mapper.Map<List<Item>>(items);
         _repositoryManager.ItemRepository.RemoveRange(mappedListItems);
         await _repositoryManager.UnitOfWork.SaveChangesAsync();
+        return true;
+        
     }
 }
